frontend/requirements.txt
matplotlib
pillow
streamlit
requests
wordcloud
pandas
plotly
